+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+--------------------
+##### Fixed
+- Loading error in Pages mode when a page has no label.
+
 [5.0.1] - 2018-12-20
 --------------------
 ##### Changed